requirements.txt
aiofiles
aiosqlite
flask
flask-cors
matplotlib
networkx
numpy
openai>=1.0.0
pandas
pyyaml
requests
tenacity
tqdm
faiss-cpu